+++ src/gcode/parser.h
@@ -4,5 +4,5 @@
 #include "gcode.h"
 
 bool lc_gcode_parser_get_value(const char *line, const char tag, float *value);
-bool lc_gcode_parser_get_command(const char **line, const lc_gcode_command_type_t command_type, uint16_t *command, bool *sub_command_existed, uint16_t *sub_command_value);
+bool lc_gcode_parser_get_command(const char **line, const lc_gcode_command_type_t command_type, uint32_t *command, bool *sub_command_existed, uint16_t *sub_command_value);
 #endif
